compare frontend: allowed scenarios configurable

+++ compare-frontend/containerInit.sh
@@ -26,11 +26,16 @@ if [[ -z "${backendPort}" ]]; then
     echo "backendPort not given"
     exit 1
 fi
+if [[ -z "${allowedScenarios}" ]]; then
+    echo "allowedScenarios not given"
+    exit 1
+fi
 
 rm "${sourceDir}"/assets/config/config.prod.json
 cp "${sourceDir}"/assets/config/config.prod.template.json "${sourceDir}"/assets/config/config.prod.json
 sed -i "s|backendUrlPlaceholder|${backendUrl}|" "${sourceDir}"/assets/config/config.prod.json
 sed -i "s|backendPortPlaceholder|${backendPort}|" "${sourceDir}"/assets/config/config.prod.json
+sed -i "s|allowedScenariosPlaceholder|${allowedScenarios}|" "${sourceDir}"/assets/config/config.prod.json
 sed -i "s|BaseHrefPlaceholder|${subPath}|" "${sourceDir}"/index.html
 
 echo "Successfully created config-file"
